requirements.txt
streamlit
langchain
langchain-openai
pillow
python-multipart
markdown